Move database connection config into store
