+++ backend/internal/store/connection.go
@@ -1,5 +1,4 @@
-// Package dbconn provides shared PostgreSQL connection configuration helpers.
-package dbconn
+package store
 
 import (
 	"fmt"
@@ -7,12 +6,12 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-// SearchPath is the backend runtime search path.
+// SearchPath is the backend runtime search path for application database pools.
 const SearchPath = "expensor,public"
 
-// ParseConfig parses a pgxpool config and applies the runtime search path used
-// by the backend application pools.
-func ParseConfig(connStr string) (*pgxpool.Config, error) {
+// ParsePoolConfig parses a pgxpool config and applies the runtime search path
+// used by the backend application pools.
+func ParsePoolConfig(connStr string) (*pgxpool.Config, error) {
 	cfg, err := pgxpool.ParseConfig(connStr)
 	if err != nil {
 		return nil, fmt.Errorf("parsing connection string: %w", err)
