+++ backend/internal/store/store.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/jackc/pgx/v5/pgxpool"
 
-	"github.com/ArionMiles/expensor/backend/internal/dbconn"
 	"github.com/ArionMiles/expensor/backend/pkg/api"
 	"github.com/ArionMiles/expensor/backend/pkg/config"
 )
@@ -42,7 +41,7 @@ func New(cfg config.Postgres, logger *slog.Logger) (*Store, error) {
 		cfg.Host, cfg.Port, cfg.User, cfg.Password, cfg.Database, cfg.SSLMode,
 	)
 
-	poolCfg, err := dbconn.ParseConfig(connStr)
+	poolCfg, err := ParsePoolConfig(connStr)
 	if err != nil {
 		return nil, fmt.Errorf("parsing store connection string: %w", err)
 	}
